+++ supervisor/process.py
@@ -1,10 +1,10 @@
-import os
-import time
 import errno
-import shlex
-import traceback
+import functools
+import os
 import signal
-from functools import total_ordering
+import shlex
+import time
+import traceback
 
 from supervisor.compat import maxint
 from supervisor.compat import as_bytes
@@ -30,7 +30,7 @@ from supervisor.datatypes import RestartUnconditionally
 
 from supervisor.socket_manager import SocketManager
 
-@total_ordering
+@functools.total_ordering
 class Subprocess(object):
 
     """A class to manage a subprocess."""
@@ -403,7 +403,8 @@ class Subprocess(object):
         self.change_state(ProcessStates.FATAL)
 
     def kill(self, sig):
-        """Send a signal to the subprocess.  This may or may not kill it.
+        """Send a signal to the subprocess with the intention to kill
+        it (to make it exit).  This may or may not actually kill it.
 
         Return None if the signal was sent, or an error message string
         if an error occurred or if the subprocess is not running.
@@ -463,14 +464,23 @@ class Subprocess(object):
             pid = -self.pid
 
         try:
-            options.kill(pid, sig)
+            try:
+                options.kill(pid, sig)
+            except OSError as exc:
+                if exc.errno == errno.ESRCH:
+                    msg = ("unable to signal %s (pid %s), it probably just exited "
+                           "on its own: %s" % (processname, self.pid, str(exc)))
+                    options.logger.debug(msg)
+                    # we could change the state here but we intentionally do
+                    # not.  we will do it during normal SIGCHLD processing.
+                    return None
+                raise
         except:
             tb = traceback.format_exc()
             msg = 'unknown problem killing %s (%s):%s' % (processname,
                                                           self.pid, tb)
             options.logger.critical(msg)
             self.change_state(ProcessStates.UNKNOWN)
-            self.pid = 0
             self.killing = False
             self.delay = 0
             return msg
@@ -502,14 +512,24 @@ class Subprocess(object):
                             ProcessStates.STOPPING)
 
         try:
-            options.kill(self.pid, sig)
+            try:
+                options.kill(self.pid, sig)
+            except OSError as exc:
+                if exc.errno == errno.ESRCH:
+                    msg = ("unable to signal %s (pid %s), it probably just now exited "
+                           "on its own: %s" % (processname, self.pid,
+                           str(exc)))
+                    options.logger.debug(msg)
+                    # we could change the state here but we intentionally do
+                    # not.  we will do it during normal SIGCHLD processing.
+                    return None
+                raise
         except:
             tb = traceback.format_exc()
             msg = 'unknown problem sending sig %s (%s):%s' % (
                                 processname, self.pid, tb)
             options.logger.critical(msg)
             self.change_state(ProcessStates.UNKNOWN)
-            self.pid = 0
             return msg
 
         return None
@@ -753,7 +773,7 @@ class FastCGISubprocess(Subprocess):
         for i in range(3, options.minfds):
             options.close_fd(i)
 
-@total_ordering
+@functools.total_ordering
 class ProcessGroupBase(object):
     def __init__(self, config):
         self.config = config