+++ supervisor/tests/test_end_to_end.py
@@ -2,8 +2,8 @@
 from __future__ import unicode_literals
 
 import os
-import sys
 import signal
+import sys
 import unittest
 import pkg_resources
 from supervisor.compat import xmlrpclib
@@ -20,7 +20,41 @@ else:
 
 class EndToEndTests(BaseTestCase):
 
+    def test_issue_291a_percent_signs_in_original_env_are_preserved(self):
+        """When an environment variable whose value contains a percent sign is
+        present in the environment before supervisord starts, the value is
+        passed to the child without the percent sign being mangled."""
+        key = "SUPERVISOR_TEST_1441B"
+        val = "foo_%s_%_%%_%%%_%2_bar"
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-291a.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        try:
+            os.environ[key] = val
+            supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+            self.addCleanup(supervisord.kill, signal.SIGINT)
+            supervisord.expect_exact(key + "=" + val)
+        finally:
+            del os.environ[key]
+
+    def test_issue_550(self):
+        """When an environment variable is set in the [supervisord] section,
+        it should be put into the environment of the subprocess."""
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-550.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact('success: print_env entered RUNNING state')
+        supervisord.expect_exact('exited: print_env (exit status 0; expected)')
+
+        args = ['-m', 'supervisor.supervisorctl', '-c', filename, 'tail -100000', 'print_env']
+        supervisorctl = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisorctl.kill, signal.SIGINT)
+        supervisorctl.expect_exact("THIS_SHOULD=BE_IN_CHILD_ENV")
+        supervisorctl.expect(pexpect.EOF)
+
     def test_issue_565(self):
+        """When a log file has Unicode characters in it, 'supervisorctl
+        tail -f name' should still work."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-565.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -36,6 +70,8 @@ class EndToEndTests(BaseTestCase):
             supervisorctl.expect_exact(line, timeout=30)
 
     def test_issue_638(self):
+        """When a process outputs something on its stdout or stderr file
+        descriptor that is not valid UTF-8, supervisord should not crash."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-638.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -53,6 +89,8 @@ class EndToEndTests(BaseTestCase):
                                      'too many start retries too quickly', timeout=60)
 
     def test_issue_663(self):
+        """When Supervisor is run on Python 3, the eventlistener protocol
+        should work."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-663.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -62,11 +100,15 @@ class EndToEndTests(BaseTestCase):
             supervisord.expect_exact('BUSY -> ACKNOWLEDGED', timeout=30)
 
     def test_issue_664(self):
+        """When a subprocess name has Unicode characters, 'supervisord'
+        should not send incomplete XML-RPC responses and 'supervisorctl
+        status' should work."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-664.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
         self.addCleanup(supervisord.kill, signal.SIGINT)
         supervisord.expect_exact('test_öäü entered RUNNING state', timeout=60)
+
         args = ['-m', 'supervisor.supervisorctl', '-c', filename, 'status']
         supervisorctl = pexpect.spawn(sys.executable, args, encoding='utf-8')
         self.addCleanup(supervisorctl.kill, signal.SIGINT)
@@ -76,6 +118,17 @@ class EndToEndTests(BaseTestCase):
         except pexpect.ExceptionPexpect:
             seen = False
         self.assertTrue(seen)
+
+    def test_issue_733(self):
+        """When a subprocess enters the FATAL state, a one-line eventlistener
+        can be used to signal supervisord to shut down."""
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-733.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact('gave up: nonexistent entered FATAL state')
+        supervisord.expect_exact('received SIGTERM indicating exit request')
+        supervisord.expect(pexpect.EOF)
 
     def test_issue_835(self):
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-835.conf')
@@ -114,7 +167,18 @@ class EndToEndTests(BaseTestCase):
             seen = False
         self.assertTrue(seen)
 
+    def test_issue_986_command_string_with_double_percent(self):
+        """A percent sign can be used in a command= string without being
+        expanded if it is escaped by a second percent sign."""
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-986.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact('dhcrelay -d -q -a %h:%p %P -i Vlan1000 192.168.0.1')
+
     def test_issue_1054(self):
+        """When run on Python 3, the 'supervisorctl avail' command
+        should work."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1054.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -129,7 +193,43 @@ class EndToEndTests(BaseTestCase):
             seen = False
         self.assertTrue(seen)
 
+    def test_issue_1170a(self):
+        """When the [supervisord] section has a variable defined in
+        environment=, that variable should be able to be used in an
+        %(ENV_x) expansion in a [program] section."""
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1170a.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact("set from [supervisord] section")
+
+    def test_issue_1170b(self):
+        """When the [supervisord] section has a variable defined in
+        environment=, and a variable by the same name is defined in
+        enviroment= of a [program] section, the one in the [program]
+        section should be used."""
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1170b.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact("set from [program] section")
+
+    def test_issue_1170c(self):
+        """When the [supervisord] section has a variable defined in
+        environment=, and a variable by the same name is defined in
+        enviroment= of an [eventlistener] section, the one in the
+        [eventlistener] section should be used."""
+        filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1170c.conf')
+        args = ['-m', 'supervisor.supervisord', '-c', filename]
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact("set from [eventlistener] section")
+
     def test_issue_1224(self):
+        """When the main log file does not need rotation (logfile_maxbyte=0)
+        then the non-rotating logger will be used to avoid an
+        IllegalSeekError in the case that the user has configured a
+        non-seekable file like /dev/stdout."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1224.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -137,6 +237,8 @@ class EndToEndTests(BaseTestCase):
         supervisord.expect_exact('cat entered RUNNING state', timeout=60)
 
     def test_issue_1231a(self):
+        """When 'supervisorctl tail -f name' is run and the log contains
+        unicode, it should not fail."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1231a.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -151,8 +253,9 @@ class EndToEndTests(BaseTestCase):
             line = '%d - hash=57d94b…381088' % i
             supervisorctl.expect_exact(line, timeout=30)
 
-
     def test_issue_1231b(self):
+        """When 'supervisorctl tail -f name' is run and the log contains
+        unicode, it should not fail."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1231b.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -185,6 +288,8 @@ class EndToEndTests(BaseTestCase):
                 break
 
     def test_issue_1231c(self):
+        """When 'supervisorctl tail -f name' is run and the log contains
+        unicode, it should not fail."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1231c.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -206,7 +311,27 @@ class EndToEndTests(BaseTestCase):
                                  timeout=30)
             supervisorctl.expect('Unicode output may fail.', timeout=30)
 
+    def test_issue_1251(self):
+        """When -? is given to supervisord or supervisorctl, help should be
+        displayed like -h does."""
+        args = ['-m', 'supervisor.supervisord', '-?']
+        supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisord.kill, signal.SIGINT)
+        supervisord.expect_exact("supervisord -- run a set of applications")
+        supervisord.expect_exact("-l/--logfile FILENAME -- use FILENAME as")
+        supervisord.expect(pexpect.EOF)
+
+        args = ['-m', 'supervisor.supervisorctl', '-?']
+        supervisorctl = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(supervisorctl.kill, signal.SIGINT)
+        supervisorctl.expect_exact("supervisorctl -- control applications")
+        supervisorctl.expect_exact("-i/--interactive -- start an interactive")
+        supervisorctl.expect(pexpect.EOF)
+
     def test_issue_1298(self):
+        """When the output of 'supervisorctl tail -f worker' is piped such as
+        'supervisor tail -f worker | grep something', 'supervisorctl' should
+        not crash."""
         filename = pkg_resources.resource_filename(__name__, 'fixtures/issue-1298.conf')
         args = ['-m', 'supervisor.supervisord', '-c', filename]
         supervisord = pexpect.spawn(sys.executable, args, encoding='utf-8')
@@ -219,6 +344,25 @@ class EndToEndTests(BaseTestCase):
         bash = pexpect.spawn('/bin/sh', ['-c', cmd], encoding='utf-8')
         self.addCleanup(bash.kill, signal.SIGINT)
         bash.expect('spewage 2', timeout=30)
+
+    def test_issue_1418_pidproxy_cmd_with_no_args(self):
+        """When pidproxy is given a command to run that has no arguments, it
+        runs that command."""
+        args = ['-m', 'supervisor.pidproxy', 'nonexistent-pidfile', "/bin/echo"]
+        pidproxy = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(pidproxy.kill, signal.SIGINT)
+        pidproxy.expect(pexpect.EOF)
+        self.assertEqual(pidproxy.before.strip(), "")
+
+    def test_issue_1418_pidproxy_cmd_with_args(self):
+        """When pidproxy is given a command to run that has arguments, it
+        runs that command."""
+        args = ['-m', 'supervisor.pidproxy', 'nonexistent-pidfile', "/bin/echo", "1", "2"]
+        pidproxy = pexpect.spawn(sys.executable, args, encoding='utf-8')
+        self.addCleanup(pidproxy.kill, signal.SIGINT)
+        pidproxy.expect(pexpect.EOF)
+        self.assertEqual(pidproxy.before.strip(), "1 2")
+
 
 def test_suite():
     return unittest.findTestCases(sys.modules[__name__])